updating home controller

+++ src/main/webapp/WEB-INF/views/display.jsp
@@ -1,9 +1,6 @@
 <%@taglib prefix="x" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
 <%@ taglib uri="http://www.springframework.org/tags" prefix="spring" %>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org"
-      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity3">
-
 <html>
 <head>
   <link rel="stylesheet" href='<x:url value="/resources/css/bootstrap.min.css"></x:url>' />
@@ -19,50 +16,41 @@
       
     </div>
     <ul class="nav navbar-nav">
-          <li><a href="cont">Contact us</a></li>
-           <sec:authorize var="loggedIn" access="isAuthenticated()" />
-          <sec:authorize access="permitAll">
-          <li><a href="${pageContext.request.contextPath}/logout"><span>Logout</span></a></li>
-		  </sec:authorize> 
-          
-         <!-- <li><a href="j_spring_security_logout">Logout</a></li> -->        
-      </ul>
+          <li><a href="hai">Home</a></li>
+          <li><a href="display">Manage Products</a></li>
+          <li><a href="">Logout</a></li>
+     </ul>
   </div>
 </nav>
  <h1>Hello... <%=session.getAttribute("loggedInUser")%></h1>
     <div class="container">
         
 <div ng-app="myApp" ng-controller="customersCtrl">
-<input type="text" class="form-control" ng-model="searchBy.watchName"/>
+<input type="text" class="form-control" ng-model="searchBy.name"/>
 <table class="table table-hover table-bordered">
   <tr>
-       <th>Product Id</th>
-       <th>Product Name</th>
-       <th>Product Description</th>
+       <th>Product id</th>
+       <th>Product name</th>
        <th>Product price</th>
+       <th>Product description</th>
       
   </tr>
   <tr ng-repeat="x in names | filter:searchBy">
     <td>{{x.id}}</td>
     <td>{{x.name}}</td>
-    <td>{{x.description}}</td>
     <td>{{x.price}}</td>
+    <td>{{x.description}}</td>
     <td>
-    
     <a href="${pageContext.servletContext.contextPath}/viewproduct?id={{x.id}}"><span>View</span></a>
-    
-    <a href="${pageContext.servletContext.contextPath}/editproduct?id={{x.id}}"><span>| Edit |</span></a>
-    <a href="${pageContext.servletContext.contextPath}/delete?id={{x.id}}"><span>Delete</span></a>
-   
+    <% if(session.getAttribute("userRole") == "ROLE_ADMIN") { %>
+    	<a href="${pageContext.servletContext.contextPath}/editproduct?id={{x.id}}"><span>| Edit |</span></a>
+    	<a href="${pageContext.servletContext.contextPath}/delete?id={{x.id}}"><span>Delete</span></a>
+    <% } %>
    </td>  
   </tr>
 </table>
-<sec:authorize access="hasRole('ROLE_ADMIN')">
-<a href="AddProduct" class="btn btn-info">addproduct</a>
-</sec:authorize>
-
+<a href="AddProduct" class="btn btn-info">Add Product</a>
 </div>
-
 <script>
 var app = angular.module('myApp', []);
 app.controller('customersCtrl', function($scope, $http) {
@@ -73,6 +61,3 @@ app.controller('customersCtrl', function($scope, $http) {
 </div>
 </body>
 </html>
-
-
-
